Tasks: Refactor loadOrCreateTokenThreshold
